hw18-spring-cloud-hystrix:правки по замечаниям
